requirements.txt

speechrecognition==3.14.1
pyttsx3==2.98
openai==1.63.2
PyAudio==0.2.14
python-dotenv==1.0.1